issue #61 : 非同期処理をoperations.tsに記述
