+++ src/edar/src/reducks/shop/reducers.ts
@@ -1,9 +1,8 @@
-import { createSlice, Dispatch, PayloadAction } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import type {
   Shop,
   Genre,
 } from '../../components/MainContent/SpecifySearchCondition/types';
-import { getGenre, getPosition, getShopList } from '../../api/api';
 
 type State = {
   position: {
@@ -137,37 +136,6 @@ const slice = createSlice({
   },
 });
 
-export function fetchPosition() {
-  return async function (dispatch: Dispatch) {
-    dispatch(setIsProcessing(true));
-    try {
-      const { latitude, longitude } = await getPosition();
-      dispatch(setPosition({ latitude, longitude }));
-    } catch (errorCode) {
-      dispatch(setPositionErrorMessage(errorCode));
-    }
-  };
-}
-
-export function fetchGenreList() {
-  return async function (dispatch: Dispatch) {
-    try {
-      dispatch(setGenreList(await getGenre()));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-}
-
-export function fetchShopList(url: string) {
-  return async function (dispatch: Dispatch) {
-    try {
-      dispatch(setShops(await getShopList(url)));
-    } catch (error) {
-      dispatch(setShopErrorMessage());
-    }
-  };
-}
 // action creatorをエクスポート
 export const {
   setPosition,
